backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiosqlite
pydantic-settings
gql[httpx]
python-dotenv
bcrypt
PyJWT
